refactor+feat: apply improvement batch from project review
Dead code / broken code:
- Drop Boundary::county() relation (County class does not exist)
- Fix Unit::scopeWithinDistance: ST_Distance_Sphere is MySQL syntax;
  PostGIS uses ST_DistanceSphere (+ regression test)
- Remove HardwareObserver + its test (never registered; superseded by
  HardwareAuditObserver)

Search quality:
- Persian/Arabic digit conversion moved into PersianNormalizer::
  normalizeForSearch so every consumer benefits
- Users page search now Persian-normalized and also matches n_code

Refactor:
- kargozini/person anonymous Livewire class extracted to
  App\Livewire\Kargozini\Person (testable, IDE-friendly). The model
  import needs an alias because the class itself is named Person.

HR dashboard:
- CSS bar charts replaced with Highcharts bar charts (per-chart max
  scaling preserved) fed via $wire.chartPayload()

Performance:
- pg_trgm GIN indexes on persons full-name expressions (both orders)
  for LIKE '%term%' searches

Housekeeping:
- PermissionSeeder now creates the admin role (tests used to each
  firstOrCreate it; two tests switched to firstOrCreate to match)
- CI test workflow clears config/route/view caches before running
