+++ app/Models/Unit.php
@@ -235,11 +235,12 @@ class Unit extends Model
      */
     public function scopeWithinDistance($query, float $lat, float $lng, float $radiusMeters)
     {
-        // MySQL/MariaDB: Use ST_GeomFromText instead of ST_Point (PostGIS-specific)
+        // PostGIS function name is ST_DistanceSphere (no underscore); the old
+        // MySQL-style ST_Distance_Sphere does not exist on this stack.
         $pointWkt = "POINT($lng $lat)";
 
         return $query->whereHas('boundary', function ($q) use ($pointWkt, $radiusMeters) {
-            $q->whereRaw('ST_Distance_Sphere(boundary, ST_GeomFromText(?, 4326)) <= ?', [$pointWkt, $radiusMeters]);
+            $q->whereRaw('ST_DistanceSphere(boundary, ST_GeomFromText(?, 4326)) <= ?', [$pointWkt, $radiusMeters]);
         });
     }
 }
